move get_binned_times_rates_pos to subpackage level utils
